ngc_inference/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ngc_inference"
version = "0.1.0"
description = "Active Inference simulations and utilities using ngclearn for variational free energy minimization"
authors = [
    {name = "NGC Inference Team", email = "ago@cs.rit.edu"},
]
readme = "README.md"
keywords = [
    'active-inference',
    'free-energy-principle', 
    'variational-inference',
    'predictive-coding',
    'ngclearn',
    'computational-neuroscience'
]
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Unix",
]
dependencies = [
    "numpy>=1.22.0",
    "scipy>=1.7.0",
    "jax>=0.4.28",
    "jaxlib>=0.4.28",
    "ngclearn>=2.0.3",
    "ngcsimlib>=1.0.1",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "loguru>=0.7.0",
    "scikit-image>=0.21.0",  # Explicit version to avoid build issues
]

[project.optional-dependencies]
dev = [
    "ipython>=8.12.0",
    "jupyter>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/NACLab/ngc-learn"
Documentation = "https://ngc-learn.readthedocs.io/"
Repository = "https://github.com/NACLab/ngc-learn"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src/ngc_inference --cov-report=term-missing --cov-report=html"
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "simulation: Full simulation tests",
    "slow: Slow running tests",
]

[tool.black]
line-length = 88
target-version = ['py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"

